Make the TCP socket wrapper into a generic async wrapper
