+++ src/sources/util/mod.rs
@@ -25,6 +25,7 @@ mod tcp_error;
 mod unix_datagram;
 #[cfg(all(unix, feature = "sources-utils-unix"))]
 mod unix_stream;
+mod wrappers;
 
 #[cfg(any(feature = "sources-http"))]
 pub use self::body_decoding::Encoding;
@@ -54,3 +55,4 @@ pub use tcp_error::TcpError;
 pub use unix_datagram::build_unix_datagram_source;
 #[cfg(all(unix, feature = "sources-utils-unix",))]
 pub use unix_stream::build_unix_stream_source;
+pub use wrappers::AfterReadExt;
